requirements.txt
sqlparse==0.4.2
gunicorn==20.1.0
Django==3.2.8
django-import-export==2.8.0
six==1.14.0
pandas==1.4.3
tmdbv3api==1.7.7
tmdbsimple==2.9.1
psycopg2==2.9.3